requirements.txt
aiogram==3.4.1
SQLAlchemy>=2.0,<3.0
asyncpg~=0.29
python-dotenv>=1.0
uvloop; sys_platform == 'linux'
alembic>=1.12
pydantic>=2.5
aiohttp>=3.8 